+++ 41_SPRINTS/2601TG/_END_SUBMISSION/Day-Spa-Animal-Pet-Purple-Br/soma subtracao divisao/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<title>soma subtracao divisao</title>
+<link rel="stylesheet" href="style.css">
+</head>
+<body>
+
+<div class="container">
+<h2>Demonstração matemática</h2>
+
+<div class="form-group">
+<label for="camponumber">Digite um número</label>
+<input type="number" id="camponumber"placeholder="Digite um número">
+
+</div>
+
+    <button id="btnSoma">Calcular Soma</button> 
+    <button id="btnSub">Calcular Subtração</button> 
+    <button id="btnDiv">Calcular Divisão</button> 
+    <button id="resetBtn">Resetar</button>
+
+<pre id="camporesposta"></pre>
+</div>
+
+<script src="script.js"></script>
+</body>
+</html>
