+++ 41_SPRINTS/2601TG/_END_SUBMISSION/Day-Spa-Animal-Pet-Purple-Br/tabuada pares/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+ <meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<title>Tabuada</title>
+<link rel="stylesheet" href="style.css">
+</head>
+<body>
+
+<div class="container">
+<h2>Tabuada Pares</h2>
+
+
+<div class="form-group">
+<label for="camponumber">Digite um número para iniciar</label>
+<input type="number" id="camponumber"placeholder="Digite um número">
+
+</div>
+
+<button id="verificarBtn">Mostrar Tabuada pares</button> 
+<button id="resetBtn">Resetar</button>
+
+<pre id="resposta"></pre>
+
+</div>
+ 
+<script src= "script.js"></script>
+</body>
+</html>
