+++ 41_SPRINTS/2601TG/_END_SUBMISSION/Day-Spa-Animal-Pet-Purple-Br/tabuada pares/style.css
@@ -1,0 +1,79 @@
+body {
+font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+background-color: antiquewhite;
+display: flex;
+justify-content: center;
+align-items: center;
+height: 100vh;
+margin: 0;
+flex-direction: column
+
+}
+
+.container {
+padding: 2rem;
+border-radius: 12px;
+box-shadow: 0 4px 15px rgba(rgba (12, 0, 0), green, blue, alpha);
+
+}
+
+h1 {
+color: blueviolet;
+font-size:0 9rem;
+
+}
+
+p{
+color: black;
+font-size: 0.9rem;
+
+}
+
+.form-group {
+text-align: left;
+margin-bottom: 1rem;
+
+}
+
+label {
+display: block;
+margin-bottom: 5px;
+font-weight: bold;
+color: cadetblue;
+
+}
+
+input {
+width: center;
+padding: 10px;
+border: 1 px solid #add;
+border-radius: 6px;
+box-sizing: border-box;
+
+}
+
+button {
+width: center;
+padding: 12px;
+background-color: chocolate;
+color: white;
+border-radius: 6px;
+cursor: pointer;
+font-size: 1rem;
+font-weight: bold;
+transition: linear;
+
+}
+
+button:hover {
+background-color: aquamarine;
+
+}
+
+mensagem {
+margin-top: 1.5rem;
+font-weight: bold;
+min-height: 1.2em;
+
+}
+
