--- 41_SPRINTS/2601TG/_END_SUBMISSION/Sorteio/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Sorteio</title>
-        <link rel="stylesheet" href="style.css">
-    </head>
-<body>
-
-    <div class="container">
-     <h2>Sorteio Aleatório</h2>
-         
-     <div class="form-group">
-<label for="campoNumber">Digite um número para iniciar</label>
-<input type="number" id="camponumber"placeholder="Digite um número">
-
-</div>
-
-<button id="guessBtn">Mostrar número</button> 
-<button id="resetBtn">Resetar</button>
-
-<p id="resultado"></p>
-
-
-     </div>
-
-     <script src= "sorteio.js"></script>
-     </body>
-     </html>
