--- 41_SPRINTS/2601TG/_END_SUBMISSION/src/index.css
@@ -1,219 +1,0 @@
-/* ==========================================================
-   1. TELA DE LOGIN BELENZINHO LOGIN
-   ========================================================== */
-.login-page {
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background-color: var(--preto-fosco);
-  padding: 20px;
-}
-
-.form-container {
-  background: white;
-  padding: 35px;           /* Reduzido um pouco (era 40px/50px) */
-  border-radius: 12px;
-  box-shadow: 0 10px 25px rgba(0,0,0,0.5);
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  width: 90%;
-  max-width: 380px;        /* Reduzido para ficar mais elegante (era 400px/480px) */
-  text-align: center;
-}
-
-.form-container h2 {
-  color: var(--preto-fosco);
-  margin-bottom: 10px;
-  border-bottom: 3px solid var(--azul-neon);
-  display: inline-block;
-  padding-bottom: 5px;
-  font-size: 1.5rem;      /* Letra um pouquinho menor */
-}
-
-/* ==========================================================
-   2. DEFINIÇÃO DE CORES E GLOBAL 
-   ========================================================== */
-:root {
-  --preto-fosco: #1a1a1a;
-  --cinza-claro: #f8f9fa;
-  --azul-neon: #00d4ff;
-  --branco: #ffffff;
-}
-
-body, html {
-  margin: 0;
-  padding: 0;
-  width: 100%;
-  background-color: var(--cinza-claro);
-  overflow-x: hidden;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-.landing-page {
-  width: 100vw; 
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.conteudo {
-  flex: 1;
-  width: 100%;
-  max-width: 1150px;      /* Reduzido de 1400px para 1150px para centralizar melhor */
-  margin: 0 auto;
-  padding: 20px;
-}
-
-/* ==========================================================
-   3. HEADER (CABEÇALHO)
-   ========================================================== */
-.header {
-  background-color: var(--preto-fosco);
-  color: white;
-  padding: 12px 5%;       /* Diminuído levemente a altura */
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  box-shadow: 0 2px 10px rgba(0,0,0,0.3);
-}
-
-/* ==========================================================
-   4. BANNER VIP
-   ========================================================== */
-.vip-banner {
-  width: 100%;
-  margin: 15px 0;
-  background: white;
-  border-radius: 12px;
-  overflow: hidden;
-  box-shadow: 0 4px 15px rgba(0,0,0,0.1);
-}
-
-.vip-img {
-  width: 100%;
-  height: 320px;           /* Reduzido de 400px para 320px */
-  object-fit: cover;
-}
-
-/* ==========================================================
-   5. GALERIA DE SERVIÇOS (VENDAS)
-   ========================================================== */
-.galeria {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 25px;               /* Diminuído o espaço entre itens */
-  padding: 30px 0;
-}
-
-.item {
-  background: white;
-  width: 290px;            /* Reduzido de 380px para 290px */
-  border-radius: 12px;
-  overflow: hidden;
-  box-shadow: 0 8px 25px rgba(0,0,0,0.15);
-  transition: 0.3s;
-}
-
-.item-img {
-  width: 100%;
-  height: 200px;           /* Reduzido para proporcionalidade */
-  object-fit: cover;
-}
-
-/* ==========================================================
-   6. ADMIN PANEL E TABELA
-   ========================================================== */
-.admin-panel {
-  width: 95%;
-  max-width: 1150px;       /* Acompanhando a redução do conteúdo */
-  margin: 0 auto;
-  padding: 30px 0;
-}
-
-.tabela-admin th, .tabela-admin td {
-  padding: 18px;           /* Reduzido de 25px para 18px */
-  font-size: 1rem;
-}
-
-/* ==========================================================
-   11. CARRINHO (DADOS ORIGINAIS)
-   ========================================================== */
-.carrinho-status {
-  position: relative;
-  font-size: 1.6rem;       /* Reduzido levemente */
-  margin-right: 20px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-}
-
-/* ==========================================================
-   12. WHATSAPP FLUTUANTE
-   ========================================================== */
-.btn-whatsapp-fixo {
-  position: fixed;
-  bottom: 20px;
-  right: 20px;
-  background-color: #25D366;
-  width: 55px;             /* Reduzido levemente de 60px para 55px */
-  height: 55px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 28px;
-  z-index: 1000;
-  box-shadow: 2px 2px 10px rgba(0,0,0,0.3);
-}
-/* ==========================================================
-   ADICIONE ESTE BLOCO NO FINAL DO SEU APP.CSS
-   ========================================================== */
-
-/* Estilo para as caixas de texto (E-mail e Senha) */
-.form-container input {
-  width: 100%;
-  padding: 12px;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  font-size: 1rem;
-  box-sizing: border-box; /* Garante que o preenchimento não aumente a caixa */
-}
-
-/* Estilo para o botão de ENTRAR */
-.btn-login {
-  background-color: #00d4ff; /* Azul Neon que você definiu */
-  color: #1a1a1a;
-  padding: 12px;
-  border: none;
-  border-radius: 6px;
-  font-weight: bold;
-  cursor: pointer;
-  font-size: 1rem;
-  transition: 0.3s;
-}
-
-.btn-login:hover {
-  background-color: #00b8e6;
-  transform: scale(1.02);
-}
-
-/* Estilo para o botão de Registrar Venda na Galeria */
-.btn-venda {
-  background-color: #1a1a1a;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  margin-top: 10px;
-  width: 100%;
-}
-
-.btn-venda:hover {
-  background-color: #333;
-}
